Updated constructor of ActionExit
I updated the constructor of ActionExit so that when an instance of that class is created the object will perform the 4 methods in the constructor: clear the drawing area, clear the edit menu, clear the status bar, and writes the word "cleaning...". The program waits 1 second before each operation. The execute member function is then performed normally.