Delete Functionality: Validation & Deletion of one component only or all the components
Validated the delete functionality to only accept deleting if there are selected component(s). Currently, it only deletes when there is ONE selected component or ALL selected components. Selecting SOME components produces an undefined behavior. I think the error lies in the fact that the deletion algorithm checks if (i == selectedComponentID), where i is the current position of the component in the ComponentList. The problem is the selectedComponentID always takes the value of i in the class ActionSelect, where i is position of the component in the CompList. However, whenever component(s) is/are deleted, the CompCount variable is reduced and the selectedComponentID will (probably) take a new value upon a new selection, yet the condition above will not match since the two "i"s are different!

We always compare the i of the ComponentList with the i of the selectedComponentID, but whenever deletion happens and you select again, the latter changes! So sometimes it deletes unselected components (because they have the selectedComponentID of a previous demanded operation) and sometimes it throws an exception and stops.